feat: 다시풀기 구현 (#69)
