+++ backend/src/main/java/com/moonspoon/moonspoon/dto/request/test/TestIncorrectProblemRequest.java
@@ -1,0 +1,10 @@
+package com.moonspoon.moonspoon.dto.request.test;
+
+import lombok.Getter;
+
+import java.util.List;
+
+@Getter
+public class TestIncorrectProblemRequest {
+    private List<Long> incorrectProblemIds;
+}
