+++ backend/src/main/java/com/moonspoon/moonspoon/problem/ProblemService.java
@@ -131,169 +131,7 @@ public class ProblemService {
         problemRepository.delete(problem);
     }
 
-    //Test logic
-    public List<TestProblemResponse> getTestProblems(Long workbookId , TestRequest dto){
 
-        Workbook workbook = validateUserAndWorkbook(workbookId);
-
-        List<Problem> problems = workbook.getProblems();
-
-        int selectCount = Math.min(dto.getProblemCount(), problems.size());
-
-        if(dto.isRandom() && !dto.getSortOrder().equals("none")){
-            //순서 정렬
-            problems = setOrderProblemList(dto.getSortOrder(), problems).subList(0, selectCount);
-            Collections.shuffle(problems);
-        }else if(dto.isRandom() && dto.getSortOrder().equals("none")){
-            Collections.shuffle(problems);
-            problems = problems.subList(0, selectCount);
-        }else if(!dto.isRandom() && !dto.getSortOrder().equals("none")){
-            //순서 정렬
-            problems = setOrderProblemList(dto.getSortOrder(), problems).subList(0, selectCount);
-        }else{
-            //순서 정렬
-            problems = setOrderProblemList("asc", problems).subList(0, selectCount);
-        }
-
-        return problems.stream()
-                .map(p -> TestProblemResponse.fromEntity(p))
-                .collect(Collectors.toList());
-    }
-
-    private List<Problem> setOrderProblemList(String order, List<Problem> problems){
-        switch(order){
-            case "asc":
-                return sortByCreateDateAsc(problems);
-            case "desc":
-                return sortByCreateDateDesc(problems);
-            case "correctRateAsc":
-                return sortByCorrectRateAsc(problems);
-            case "correctRateDesc":
-                return sortByCorrectRateDesc(problems);
-            default:
-                return problems;
-        }
-    }
-
-    private List<Problem> sortByCorrectRateAsc(List<Problem> problems){
-        return problems.stream()
-                .sorted(Comparator.comparingDouble(Problem::getCorrectRate))
-                .collect(Collectors.toList());
-    }
-
-    private List<Problem> sortByCorrectRateDesc(List<Problem> problems){
-        return problems.stream()
-                .sorted(Comparator.comparingDouble(Problem::getCorrectRate).reversed())
-                .collect(Collectors.toList());
-    }
-
-    private List<Problem> sortByCreateDateAsc(List<Problem> problems){
-        return problems.stream()
-                .sorted(Comparator.comparing(Problem::getCreateDate))
-                .collect(Collectors.toList());
-    }
-
-    private List<Problem> sortByCreateDateDesc(List<Problem> problems){
-        return problems.stream()
-                .sorted(Comparator.comparing(Problem::getCreateDate).reversed())
-                .collect(Collectors.toList());
-    }
-
-    public List<TestResultResponse> getTestResultProblem(Long workbookId){
-        //검증 로직
-        validateUserAndWorkbook(workbookId);
-        String username = SecurityContextHolder.getContext().getAuthentication().getName();
-        List<TestResultRequest> dto = storedLists.get(username);
-        List<Long> problemIds = dto.stream().map(TestResultRequest::getId).collect(Collectors.toList());
-
-        List<Problem> problems = problemRepository.findByIdList(problemIds);
-        Map<Long, Problem> problemMap = problems.stream()
-                .collect(Collectors.toMap(Problem::getId, problem -> problem));
-
-        List<TestResultResponse> responses =  dto.stream()
-                .map(d -> {
-                    Problem problem = problemMap.get(d.getId());
-                    TestResultResponse res = TestResultResponse.fromEntity(problem);
-                    res.setInput(d.getInput());
-                    res.setResult(compareStrings(d.getInput(), problem.getSolution()));
-                    return res;
-                })
-                .collect(Collectors.toList());
-
-        return responses;
-
-    }
-
-    private String compareStrings(String input, String sol){
-        String inputData = input.replaceAll("\\s", "").toLowerCase();
-        String solution = sol.replaceAll("\\s", "").toLowerCase();
-        if(inputData.equals(solution)){
-            return "correct";
-        }else{
-            return "incorrect";
-        }
-    }
-
-    @Transactional
-    public TestResultSubmitResponse testResultSubmit(Long workbookId, List<TestResultSubmitRequest> dto){
-        //검증 로직
-        validateUserAndWorkbook(workbookId);
-        //계산 및 검증로직
-        int size = dto.size();
-        int correctCount = 0;
-        for(int i = 0; i < dto.size(); i++){
-            if(dto.get(i).getResult().equals("correct")){
-                correctCount++;
-            }
-        }
-        //내 문제집 테스트 횟수 증가
-        User user = getCurrentUser();
-        user.addWorkbookTestCount();
-
-        TestResultSubmitResponse response = TestResultSubmitResponse.builder()
-                .correctCount(correctCount)
-                .incorrectCount(size-correctCount)
-                .score(((double)correctCount/size)*100)
-                .problems(dto)
-                .build();
-
-        // 벌크 연산
-        updateCorrectRate(dto);
-
-        return response;
-    }
-
-    private void updateCorrectRate(List<TestResultSubmitRequest> dto){
-        List<Long> correctProblemIds = new ArrayList<>();
-        List<Long> incorrectProblemIds = new ArrayList<>();
-
-        for (TestResultSubmitRequest result : dto) {
-            if ("correct".equals(result.getResult())) {
-                correctProblemIds.add(result.getId());
-            } else {
-                incorrectProblemIds.add(result.getId());
-            }
-        }
-        if (!correctProblemIds.isEmpty()) {
-            em.createQuery("UPDATE Problem p SET p.correctCount = p.correctCount + 1 WHERE p.id IN :ids")
-                    .setParameter("ids", correctProblemIds)
-                    .executeUpdate();
-        }
-
-        if (!incorrectProblemIds.isEmpty()) {
-            em.createQuery("UPDATE Problem p SET p.incorrectCount = p.incorrectCount + 1 WHERE p.id IN :ids")
-                    .setParameter("ids", incorrectProblemIds)
-                    .executeUpdate();
-        }
-
-    }
-
-    public void storeInputData(Long workbookId ,List<TestResultRequest> listDto){
-        //검증 로직
-        validateUserAndWorkbook(workbookId);
-        String username = SecurityContextHolder.getContext().getAuthentication().getName();
-        this.storedLists.put(username, listDto);
-    }
 
     private User getCurrentUser(){
         String username = SecurityContextHolder.getContext().getAuthentication().getName();
