+++ frontend/src/router/index.js
@@ -19,6 +19,7 @@ import SharedWorkbookDetail from "@/components/sharedWorkbook/SharedWorkbookDeta
 import SharedProblemTest from "@/components/sharedTest/SharedProblemTest.vue";
 import SharedScoringTest from "@/components/sharedTest/SharedScoringTest.vue";
 import Profile from "@/components/Profile.vue";
+import IncorrectProblemTest from "@/components/test/IncorrectProblemTest.vue";
 
 
 Vue.use(VueRouter);
@@ -107,6 +108,11 @@ const routes = [
       path: '/profile',
       name: 'Profile',
       component: Profile
+    },
+    {
+      path: '/incorrectProblemTest',
+      name: 'IncorrectProblemTest',
+      component: IncorrectProblemTest
     },
     {
         path: '/test',
